requirements.txt
flask==3.0.0
qc-procrustes>=1.0.0
numpy==1.26.2
pandas==2.1.4
openpyxl>=3.0.9
werkzeug>=3.0.0
gunicorn==21.2.0
gevent==23.9.1
redis==5.0.1
celery==5.3.6
flower==2.0.1
